requirements.txt
# Core Deep Learning Frameworks
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Face Detection & Recognition
# Note: retinaface removed due to TensorFlow version conflict (requires TF 2.5.0)
# Note: insightface removed due to C++ build tools requirement
# Use facenet-pytorch for face detection and recognition
facenet-pytorch>=2.5.3

# Scientific Computing
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Image Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Additional utilities
colorama>=0.4.6

# Serial Communication (for Proteus integration)
pyserial>=3.5
